build: Move static functions and remove parameters
